mk2/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pyngrok>=5.0.5
torch==2.1.1
torchvision>=0.10.0
opencv-python==4.8.1.78
numpy==1.26.2
Pillow==10.1.0
ultralytics==8.0.196
transformers==4.35.2
python-dotenv>=0.19.0
pydantic>=1.8.2
python-jose==3.3.0 